extract manchester ontology doc into mod
